% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvlasso.R
\name{cv.lasso}
\alias{cv.lasso}
\title{Computes K-fold cross-validated mean squared error for lasso}
\usage{
cv.lasso(X, Y, group = 1:ncol(X), weights = rep(1, max(group)), type,
  K = 10L, minlbd, maxlbd, num.lbdseq = 100L, parallel = FALSE,
  ncores = 2L, plot.it = FALSE, verbose = FALSE)
}
\arguments{
\item{X}{predictor matrix of size \code{p} x \code{n}.}

\item{Y}{response vector of length \code{n}.}

\item{group}{p x 1 vector of consecutive integers. The number of groups
should be same as max(group).}

\item{weights}{weight vector of length \code{max(group)}.
Weight term for each group. Default is \code{rep(1, max(group))}.}

\item{type}{type of penalty, either to be "lasso", "grlasso", "slasso" or "sgrlasso".}

\item{K}{integer. number of folds}

\item{minlbd}{numeric. Minumum value of the lambda sequence.}

\item{maxlbd}{numeric. Maximum value of the lambda sequence.}

\item{num.lbdseq}{integer. Number of the lambda sequence.}

\item{parallel}{logical. If \code{parallel = TRUE}, uses parallelization.
Default is \code{parallel = FALSE}.}

\item{ncores}{integer. The number of cores to use for the parallelization.}

\item{plot.it}{logical. If ture, plot the squared error curve}

\item{verbose}{verbose}
}
\value{
\item{lbd.min}{a value of lambda such gives a minimum squared error.}

\item{lbd.1se}{a largest lambda within 1 std. from \code{lbd.min}.}

\item{lbd.seq}{lambda sequence.}

\item{cv}{mean squared error at each lambda value.}

\item{cvsd}{the standard deviaion of cv.}
}
\description{
Computes K-fold cross-validated mean squared error
to propose a lambda value for lasso, group lasso, scaled lasso or scaled
group lasso
}
\examples{
set.seed(123)
n <- 30
p <- 50
group <- rep(1:(p/10),each=10)
weights <- rep(1, max(group))
X <- matrix(rnorm(n*p),n)
truebeta <- c(rep(1,5),rep(0,p-5))
Y <- X\%*\%truebeta + rnorm(n)

cv.lasso(X,Y,group,weights,K=5,type="sgrlasso",num.lbdseq=10,plot.it=TRUE)
cv.lasso(X,Y,group,weights,K=10,type="grlasso",num.lbdseq=100,plot.it=TRUE)
}
