% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectSampler.R
\name{DirectSampler}
\alias{DirectSampler}
\title{This function directly draw samples with bootstrap using gglasso package.}
\usage{
DirectSampler(X, Y, coeff, sigma2, lbd, weights, group, niter = 2000,
  futype = "normal", center = FALSE, parallel = FALSE, ncores,
  verbose = FALSE)
}
\arguments{
\item{X}{n x p matrix of predictors.}

\item{Y}{n x 1 vector of response. Not needed for futype="bootstrap".}

\item{coeff}{n x 1 vector of estimates of true coefficient.}

\item{sigma2}{variance of error term. Not needed for futype="bootstrap".}

\item{lbd}{penalty term of lasso. See the loss function given at details.}

\item{weights}{Weight term for each group. Default is rep(1, max(group)).}

\item{group}{p x 1 vector of consecutive integers. The number of groups should be same as max(group).}

\item{niter}{The number of iterations.}

\item{futype}{Method for resampling errors. Either to be "noraml" or "bootstrap". Default is "normal".}

\item{center}{Whether to center Y values. Default is FALSE.}

\item{parallel}{Whether to parallelize the code. Default is FALSE.}

\item{ncores}{The number of cores to use for the parallelization. If missing, it uses maximum number of cores.}

\item{verbose}{Whether to show the process. Default is FALSE. Only works when parallel=FALSE.}
}
\value{
\describe{
  \item{beta}{coefficient matrix of size N x p.}
  \item{subgrad}{subgradient matrix of size N x p.}
 }
}
\description{
This function directly draw samples with bootstrap using gglasso package.
}
\details{
If futype="normal", it generate
}
\examples{
n <- 10
p <- 30
lbd <-  .5
niter <-  10
group <- rep(1:(p/10), each=10)
Weights <- rep(1,p/10)
x <- matrix(rnorm(n*p), n)
DirectSampler(X = x, coeff = rep(0,p), sigma2=1, lbd=.5, weights=Weights, group=group,N=niter, parallel=FALSE)
DirectSampler(X = x, coeff = rep(0,p), sigma2=1, lbd=.5, weights=Weights, group=group,N=niter, parallel=TRUE)
}
