% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHInference.R
\name{Lasso.MHLS}
\alias{Lasso.MHLS}
\title{lasso / group lasso estimator}
\usage{
Lasso.MHLS(X, Y, lbd = 0.37, weights = rep(1, max(group)),
  group = 1:ncol(X))
}
\arguments{
\item{X}{n x p matrix of predictors.}

\item{Y}{n x 1 vector of response.}

\item{lbd}{penalty term of lasso. See the loss function given below for more
details.}

\item{weights}{Weight term for each group. Default is
\code{rep(1, max(group))}.}

\item{group}{p x 1 vector of consecutive integers. The number of groups
should be same as max(group).}
}
\value{
\item{B0}{a vector of coefficient estimator.}

\item{S0}{a vector of subgradient.}

\item{lbd, weights, grouop}{same as input arguments.}
}
\description{
provides lasso / group lasso solutio;
coefficient-estimate and subgradient.
}
\details{
Using gglasso package, provide lasso / group lasso solution along with
subgradient. The loss function for group lasso is
\deqn{L(\beta) = ||y-X\beta||^2 / (2n) + \lambda \sum_j ||\beta_(j)||,}
where (j) is the index set of j-th group. If the size of the group is 1 for
every group, it becomes lasso loss function.
}
\examples{
set.seed(123)
n <- 50
p <- 10
X <- matrix(rnorm(n*p),n)
Y <- X \%*\% c(1,1,rep(0,p-2)) + rnorm(n)
#
# lasso
#
Lasso.MHLS(X = X,Y = Y,lbd = .5)
#
# group lasso
#
Lasso.MHLS(X = X,Y = Y,lbd = .5,weights = rep(1,2),group=rep(1:2,each=5))
}
