% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBsampler.R
\name{PB.CI}
\alias{PB.CI}
\title{Provide \code{(1-alpha)\%} confidence interval of each coefficients}
\usage{
PB.CI(object, alpha = 0.05, method = "debias", parallel = FALSE,
  ncores = 2L)
}
\arguments{
\item{object}{Bootstrap samples of class \code{PB} from \code{PBsampler}}

\item{alpha}{significance level.}

\item{method}{bias-correction method. Either to be "none" or "debias".}

\item{parallel}{Logical. If \code{TRUE}, use parallelization.}

\item{ncores}{Integer. The number of cores to use for the parallelization.}
}
\value{
\code{(1-alpha)\%} confidence interval of each coefficients
}
\description{
Using samples drawn by \code{\link{PBsampler}}, compute
\code{(1-alpha)\%} confidence interval of each coefficients.
}
\details{
If \code{method==none}, \code{\link{PB.CI}} simply compute
the quantile of the sampled coefficients. If \code{method==debias}, we use
debiased estimator to compute confidence interval.
}
\examples{
set.seed(1234)
n <- 40
p <- 50
Niter <-  10
Group <- rep(1:(p/10), each = 10)
Weights <- rep(1, p/10)
X <- matrix(rnorm(n*p), n)
object <- PBsampler(X,c(1,1,rep(0,p-2)),1,.5,niter=100,type="lasso")
parallel <- (.Platform$OS.type != "windows")
PB.CI(object = object, alpha = .05, method = "none")

}
\references{
Zhang, C., Zhang, S. (2014) Confidence intervals for low dimensional
parameters in high dimensional linear models. Journal of the Royal
Statistical Society: Series B 76, 217–242.
}
