% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHLS.R
\name{plot.MHLS}
\alias{plot.MHLS}
\title{Plotting Metropolis-Hastings sampler outputs}
\usage{
\method{plot}{MHLS}(x, index = 1:ncol(x$beta), skipS = FALSE, ...)
}
\arguments{
\item{x}{an object of class "MHLS", which is a result of \code{\link{MHLS}}.}

\item{index}{an index of covariates that one can plot with.}

\item{skipS}{logical. If \code{TRUE}, plot beta only.}

\item{...}{...    further arguments passed to or from other methods.}
}
\description{
povides six plots for each covariates index;
 histogram, path plot and acf plot for beta and subgradient.
}
\details{
\code{plot.MHLS} provides summary plots of sampled beta and subgradient.
 The first column provides histogram of beta and subgradient, while the second
 and the third column provide path plot and acf plot, respectively.
 If \code{skipS = TRUE}, it will provide summary plots for beta only.
}
\examples{
#' set.seed(123)
n <- 10
p <- 5
X <- matrix(rnorm(n * p), n)
Y <- X \%*\% rep(1, p) + rnorm(n)
sigma2 <- 1
lbd <- .37
weights <- rep(1, p)
LassoResult <- Lasso.MHLS(X = X, Y = Y, lbd = lbd, weights = weights)
B0 <- LassoResult$B0
S0 <- LassoResult$S0
plot(MHLS(X = X, pointEstimate = rep(0, p), sig2 = 1, lbd = 1, group = 1:p,
     weights = weights, B0 = B0, S0 = S0, niter = 50, burnin = 0,
     type = "coeff"))
}
