% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHLS.R
\name{summary.MHLS}
\alias{summary.MHLS}
\title{Summarizing Metropolis-Hastings sampler outputs}
\usage{
\method{summary}{MHLS}(object)
}
\arguments{
\item{object}{an object of class "MHLS", which is a result of \code{\link{MHLS}}.}
}
\value{
mean, median, s.d., 2.5% quantile and 97.5% quantile for each sampled beta and subgradient.
}
\description{
summary method for class "MHLS"
}
\details{
This function provides a summary of each sampled beta and subgradient.
}
\examples{
#' set.seed(123)
n <- 10
p <- 5
X <- matrix(rnorm(n * p), n)
Y <- X \%*\% rep(1, p) + rnorm(n)
sigma2 <- 1
lbd <- .37
weights <- rep(1, p)
LassoResult <- Lasso.MHLS(X = X, Y = Y, lbd = lbd, weights = weights)
B0 <- LassoResult$B0
S0 <- LassoResult$S0
summary(MHLS(X = X, pointEstimate = rep(0, p), sig2 = 1, lbd = 1, group = 1:p,
     weights = weights, B0 = B0, S0 = S0, niter = 50, burnin = 0,
     type = "coeff"))
}
