% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportanceSampler.R
\name{hdIS}
\alias{hdIS}
\title{Computing importance weights under high-dimensional setting}
\usage{
hdIS(X, coefftarget, sig2target, lbdtarget, coeffprop1, sig2prop1, lbdprop1,
  coeffprop2, sig2prop2, lbdprop2, proposalsample, group, weights = rep(1,
  length(unique(group))), TsA.method = "default", log = TRUE,
  parallel = FALSE, ncores)
}
\arguments{
\item{X}{Predictor matrix of size n x p, where n is the number of
observations and p is number of covariates.}

\item{coefftarget, sig2target, lbdtarget}{Parameter of target distribution.
(coefficient estimate, estimated variance of error, lambda)}

\item{coeffprop1, sig2prop1, lbdprop1}{Parameter of proposal distribution.
(coefficient estimate, estimated variance of error, lambda).}

\item{coeffprop2, sig2prop2, lbdprop2}{Parameter of mixture proposal
distribution. (coefficient estimate, estimated variance of error, lambda).
Only needed for group lasso under mixture distribution.
See Zhou and Min(2016) for more details.}

\item{group}{Prespecifed group structure for group lasso. p x 1 vector.
Use the same integer if covariates are in the same group. See example for
more detail.}

\item{weights}{Weight term for each group. Default is
rep(1, length(unique(group))).}

\item{TsA.method}{Way to construct T(eta(s),A) matrix. See Zhou and Min(2016)
for more detail.}

\item{log}{If true, importance weight is computed in log scale.}

\item{parallel}{logical. If true, parallelize the computation.}

\item{ncores}{number of CPU cores to use for parallelization.}

\item{propsalsample}{Samples drawn from proposal distribution. The samples
can be easily drawn by using \code{\link{DirectSampler}} with relevant choice
of \code{coeff, sigma2, lbd}.}
}
\value{
\code{hdIS} returns importance weights of the proposed sample.
}
\description{
\code{hdIS} is used to computes importance weights using samples
drawn by \code{\link{DirectSampler}}. For group lasso, we provide the option
to use mixture distribution as a proposal distribution. See the examples
below for more details
}
\details{
Computes importance weights which is defined as \deqn{\frac{target
 density}{proposal density}}, when the samples is drawn from proposal
 distribution with (coeffprop, sig2prop, lbdprop) while the parameter of
 target distribution is (coefftarget, sig2target, lbdtarget).
}
\examples{
n <- 10
p <- 30
lbd <-  .5
niter <-  10
group <- rep(1:(p/10), each=10)
Weights <- rep(1,p/10)
x <- matrix(rnorm(n*p), n)
DirectSampler(X = x, coefftarget = rep(0,p), sig2target=1, lbd=.5,
weights=Weights, group=group,N=niter, parallel=FALSE)
DirectSampler(X = x, coefftarget = rep(0,p), sig2target=1, lbd=.5,
weights=Weights, group=group,N=niter, parallel=TRUE)
}
