% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHInference.R
\name{Lasso.MHLS}
\alias{Lasso.MHLS}
\title{Compute lasso estimator}
\usage{
Lasso.MHLS(X, Y, type = "lasso", lbd, group = 1:ncol(X), weights = rep(1,
  max(group)), verbose = FALSE, ...)
}
\arguments{
\item{X}{n x p matrix of predictors.}

\item{Y}{n x 1 vector of response.}

\item{type}{type of penalty, either to be "lasso", "grlasso", "slasso" or "sgrlasso".}

\item{lbd}{penalty term of lasso. By letting this argument to \code{"cv.1se"} or
\code{"cv.min"}, users can have the cross-validated lambda that gives either minimum
squared error or that is within 1 std error bound.}

\item{group}{p x 1 vector of consecutive integers. The number of groups
should be same as max(group).}

\item{weights}{Weight term for each group. Default is
\code{rep(1, max(group))}.}

\item{verbose}{verbose for type = "slasso" or "sgrlasso".}

\item{...}{Auxiliary arguments for \code{lbd="cv.min", lbd="cv.1se"}.
Check \code{\link{cv.lasso}} for more details.}
}
\value{
\item{B0}{a vector of coefficient estimator.}

\item{S0}{a vector of subgradient.}

\item{lbd, weights, group}{same as input arguments.}
}
\description{
provides lasso / group lasso / scaled lasso / scaled grouop lasso solution;
coefficient-estimate and subgradient along with sigma-estimate
if scaled lasso or scaled group lasso.
}
\details{
Using gglasso package, provide lasso / group lasso solution along with
subgradient. The loss function for group lasso is
\deqn{L(\beta) = ||y-X\beta||^2 / (2n) + \lambda \sum_j ||\beta_(j)||,}
where (j) is the index set of j-th group. If the size of the group is 1 for
every group, it becomes lasso loss function.
}
\examples{
set.seed(123)
n <- 50
p <- 10
X <- matrix(rnorm(n*p),n)
Y <- X \%*\% c(1,1,rep(0,p-2)) + rnorm(n)
#
# lasso
#
Lasso.MHLS(X = X,Y = Y,lbd = .5)
#
# group lasso
#
Lasso.MHLS(X = X,Y = Y, type="grlasso", lbd = .5,weights = rep(1,2),group=rep(1:2,each=5))
}
