% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAlasso.R
\name{MHLS}
\alias{MHLS}
\title{Metropolis-Hastings sampler using estimator augmentation in lasso estimator.}
\usage{
MHLS(X, ...)
}
\arguments{
\item{X}{n x p matrix of predictors.}

\item{pointEstimate}{n x 1 vector of estimates of true coefficient.}

\item{sig2}{variance of error term.}

\item{weights}{Weight term for each coefficient. Default is rep(1, p).}

\item{lbd}{penalty term of lasso. See the loss function given at details.}

\item{group}{p x 1 vector of consecutive integers.
The number of groups should be same as max(group).}

\item{niter}{The number of iterations. Default value is 2000.}

\item{burnin}{The length of burin-in periods. Default value is 0.}

\item{B0}{Initial value of coefficients of length p.
Make sure to set values to zero for inactive set and non-zero for active set.}

\item{S0}{Initial value of subgradients. Has to satisfy certain conditions
for subgradients. If not given, it will be generated in defauly way}

\item{tau}{p x 1 vector. Variance parameter for proposal
distribution of coefficients.}

\item{indexweights}{p x 1 vector. Weights for updating active/inactive Index.}

\item{FlipSA}{The parameter that is needed for the high-dimensional setting.
Has to be a subset of active set, A. If the index is not listed in FlipSA,
the sign of coefficients which corresponds to the index will be fixed.
The default is \code{FlipSA=A}.}

\item{SFindex}{Subgradient index for the free coordinate.}

\item{randomSFindex}{logical. If \code{true}, resample \code{SFindex} in every
\code{updateSFindex} number.}

\item{updateSFindex}{Specifies how many iterations will be done without
updating the SFindex.}

\item{verbose}{verbose}

\item{updateS.itv}{Update S in every updateS.itv iterations. Set this value larger than niter if one wants to skip updating subgradients.}
}
\value{
\describe{
  \item{beta}{coefficient matrix of size N x p.}
  \item{subgrad}{subgradient matrix of size N x p.}
 }
}
\description{
Metropolis-Hastings sampler using estimator augmentation in lasso estimator.
}
\details{
If futype="normal", it generate
}
\examples{
n <- 10
p <- 30
lbd <- .5
niter <-  10
group <- rep(1:(p/10), each=10)
Weights <- rep(1,p/10)
x <- matrix(rnorm(n*p), n)
DirectSampler(X = x, pointEstimate = rep(0,p), sig2=1, lbd=.5, weights=Weights,
group=group,N=niter, parallel=FALSE)
DirectSampler(X = x, pointEstimate = rep(0,p), sig2=1, lbd=.5, weights=Weights,
group=group,N=niter, parallel=TRUE)
}
