% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAlasso.R
\name{plot.MHLS}
\alias{plot.MHLS}
\title{Plotting Metropolis-Hastings sampler outputs}
\usage{
\method{plot}{MHLS}(object, A = 1:ncol(object$beta), skipS = FALSE, ...)
}
\arguments{
\item{object}{an object of class "MHLS", which is a result of \code{\link{lm}}.}

\item{A}{an index of covariates that one can plot with.}

\item{skipS}{logical. If \code{TRUE}, plot betas only.}
}
\value{
this function gives mean, median, s.d., 2.5% quantile and 97.5% quantile for each sampled beta and subgradient.
}
\description{
For each index, this provides six plots;
 histogram, path plot and acf plot for beta and subgradient.
}
\details{
\code{summary.MHLS} provides a summary of each sampled beta and subgradient.
}
\examples{
#' set.seed(123)
n <- 10
p <- 5
X <- matrix(rnorm(n * p), n)
Y <- X \%*\% rep(1, p) + rnorm(n)
sigma2 <- 1
lbd <- .37
weights <- rep(1, p)
LassoResult <- Lasso.MHLS(X = X, Y = Y, lbd = lbd, weights = weights)
B0 <- LassoResult$B0
S0 <- LassoResult$S0
summary(MHLS(X = X, pointEstimate = rep(0, p), sig2 = 1, lbd = 1, group = 1:p,
     weights = weights, B0 = B0, S0 = S0, niter = 50, burnin = 0,
     type = "coeff"))
}
