% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHInference.R
\name{Postinference.MHLS}
\alias{Postinference.MHLS}
\title{Post-inference for lasso estimator}
\usage{
Postinference.MHLS(X, Y, B0, S0, lbd, weights = rep(1, length(B0)),
  tau = rep(1, sum(B0 != 0)), sig2.hat, alpha = 0.05, nChain = 10,
  niterPerChain = 500, parallel = FALSE, ncores = 2L,
  printSamples = FALSE, ...)
}
\arguments{
\item{X}{predictors matrix.}

\item{Y}{response vector.}

\item{B0}{Lasso estimator.}

\item{S0}{Subgradient which corresponds to B0.}

\item{lbd}{penalty term of lasso. See the loss function given below for
more details.}

\item{weights}{Weight term for each group. Default is
\code{rep(1, max(group))}.}

\item{tau}{numeric vector. Standard deviaion of proposal distribution
 for each active beta.
Adjust the value to get relevant level of acceptance rate.}

\item{sig2.hat}{variance of error term.}

\item{alpha}{confidence level for confidence interval.}

\item{nChain}{The number of chains each from different pluginbeta.}

\item{niterPerChain}{The number of iteration per chain.}

\item{parallel}{Whether to parallelize the code. Default is \code{FALSE}.}

\item{ncores}{The number of cores to use for the parallelization. If missing,
it uses maximum number of cores.}

\item{printSamples}{Boolean value. If true, print Metropolis-Hastings samples.}

\item{...}{auxiliary \code{link{MHLS}} arguments.}
}
\value{
\item{MHsamples}{a list of a class MHLS.}

\item{confidenceInterval}{(1-alpha) confidence interval
of each active coefficient.}
}
\description{
Provides Confidence intervals for the set of active coefficients
from lasso estimator using Metropolis-Hastings sampler.
}
\details{
We provide Post-selection inference for lasso estimator.
Using Metropolis Hastings Sampler with multiple chanin, \code{(1-alpha)}
confidence interval for each active coefficients is generated.
Set \code{printSamples=TRUE} if one wants to check the samples.
Check the acceptance rate and adjust tau accordingly. Desirable level of
acceptance rate for beta is \code{0.30 +- 0.15}. We recommend to set
nChain >= 10 and niterPerChain >=500.
}
\examples{
set.seed(123)
n <- 5
p <- 10
X <- matrix(rnorm(n*p),n)
Y <- X \%*\% rep(1,p) + rnorm(n)
sig2 <- 1
lbd <- .37
weights <- rep(1,p)
group <- 1:p
LassoResult <- Lasso.MHLS(X = X,Y = Y,lbd = lbd,group=group,
weights = weights)
B0 <- LassoResult$B0
S0 <- LassoResult$S0
Postinference.MHLS(X, Y, B0, S0, lbd, weights, tau=rep(1, sum(B0!=0)),
sig2.hat=1, alpha=.05, nChain=3, niterPerChain=20, parallel=TRUE)
Postinference.MHLS(X, Y, B0, S0, lbd, weights, tau=rep(1, sum(B0!=0)),
sig2.hat=1, alpha=.05, nChain=3, niterPerChain=20,
parallel=TRUE, printSamples=TRUE)
}
