% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHInference.R
\name{postInference.MHLS}
\alias{postInference.MHLS}
\title{Post-selection individual inference with lasso estimator}
\usage{
postInference.MHLS(LassoEst, Ctype = "CI", X, Y, sig2.hat, tau = rep(1,
  ncol(X)), alpha = 0.05, MHsamples, target = which(LassoEst$B0 != 0),
  nChain = 10, method, niterPerChain = 500, parallel = FALSE,
  ncores = 2L, returnSamples = FALSE, ...)
}
\arguments{
\item{LassoEst}{The result from \code{\link{lassoFit}} function with \code{type="lasso"}.}

\item{Ctype}{either \code{"CI"} or \code{"CS"} which represent confidence intervals
and confidence sets, respectively. If \code{"CI"}, confidence intervals for all active
coefficients are generated. If \code{"CS"}, \code{target} argument needs to be specified.}

\item{X}{predictor matrix.}

\item{Y}{response vector.}

\item{sig2.hat}{variance of error term.}

\item{tau}{numeric vector. Standard deviation of proposal distribution
for each beta. Adjust the value to get relevant level of acceptance rate.
Default is \code{rep(1, ncol(X))}.}

\item{alpha}{confidence level for confidence interval.}

\item{MHsamples}{optional argument. MHsamples from \code{\link{postInference.MHLS}}.
If MHsamples is supplied, MH sampling step is omitted. See the example for the detail.}

\item{target}{target active variables of which one wants to generate confidence set. Needs
to be a subset of active set. See the example for the detail.}

\item{nChain}{the number of chains. For each chain, different plug-in beta will be generated
from its confidence region.}

\item{method}{Type of robust method. Users can choose either \code{"coeff"} or \code{"mu"}.}

\item{niterPerChain}{the number of iterations per chain.}

\item{parallel}{logical. If \code{parallel = TRUE}, uses parallelization.
Default is \code{parallel = FALSE}.}

\item{ncores}{integer. The number of cores to use for parallelization.}

\item{returnSamples}{logical. If \code{returnSamples = TRUE}, print Metropolis-Hastings samples.
If \code{MHsamples} is supplied, \code{returnSamples = FALSE} is forced.}

\item{...}{auxiliary \code{\link{MHLS}} arguments.}
}
\value{
\item{MHsamples}{a list of class MHLS.}

\item{confidenceInterval}{(1-alpha) confidence interval
for each active coefficient.}
}
\description{
Provides confidence intervals for the set of active coefficients
of lasso using Metropolis-Hastings sampler.
}
\details{
This function provides post-selection inference for the active coefficients selected by lasso.
Uses Metropolis-Hastings sampler with multiple chains to draw from the
distribution under a fixed active set and generates \code{(1-alpha)}
confidence interval for each active coefficients.
Set \code{returnSamples = TRUE} to check the Metropolis-Hastings samples.
Check the acceptance rate and adjust \code{tau} accordingly.
We recommend to set \code{nChain >= 10} and \code{niterPerChain >= 500}.
}
\examples{
set.seed(123)
n <- 6
p <- 10
X <- matrix(rnorm(n*p),n)
Y <- X \%*\% rep(1,p) + rnorm(n)
sig2 <- 1
lbd <- .37
weights <- rep(1,p)
LassoEst <- lassoFit(X = X, Y = Y, type = "lasso", lbd = lbd, weights = weights)
parallel <- (.Platform$OS.type != "windows")
P1 <- postInference.MHLS(LassoEst= LassoEst, X = X, Y = Y, sig2.hat = 1, alpha = .05,
nChain = 3, niterPerChain = 20, method = "coeff", parallel = parallel, returnSamples = TRUE)
P1
postInference.MHLS(LassoEst= LassoEst, MHsamples = P1$MHsamples,
                   Ctype = "CI", X = X, Y = Y, method = "coeff")
postInference.MHLS(LassoEst= LassoEst, MHsamples = P1$MHsamples,
                   Ctype = "CS", X = X, Y = Y, method = "coeff")
}
